Resources/python/api/requirements.txt
# Server
fastapi==0.115.5
uvicorn==0.32.1


# Configuration
pydantic==2.9.2
pydantic-settings==2.6.1
python-dotenv==1.0.0

# RDF and Triple Store
rdflib==7.1.1
requests==2.32.3